Show an icon on loaded filters.
We use a QStyledItemDelegate to show the icon at the right side.
Currently we only use a single one, though soon we will show an icon the
same way on a few other elements. That's the reason for having an array
of delegates, even though currently there's only one element.